app/backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
httpx>=0.27.0
pydantic>=2.0.0
py7zr>=0.22.0

# ===== Static Analysis Scanners =====
python-magic>=0.4.27
oletools>=0.60
pdfid>=1.1.0
pefile>=2023.2.7
lief>=0.15.0
ppdeep>=20200515
androguard>=4.1
flare-capa>=7.0
pytest>=7.0.0
